gcp_/14-06-2023/requirements.txt
Flask==3.0.0
requests
google-cloud-firestore
google-cloud-pubsub
pyyaml